Crate serde_derive_internals

Modules

ast

A Serde ast, parsed from the Syn ast and ready to generate Rust code.

attr

Structs

Ctxt

A type to collect errors together and format them.

Enums

Derive

Functions

replace_receiver

ungroup