Struct serde_derive_internals::ast::Repr

pub struct Repr {
    pub int_repr: Option<&'static str>,
    pub c_repr: bool,
    pub other_repr: bool,
}

Fields
int_repr: Option<&'static str> c_repr: bool other_repr: bool
Methods

impl Repr

pub fn get_stable_rust_enum_layout(&self) -> Option<&'static str>

Gives the int type to use for the repr(int) enum layout

pub fn get_stable_c_enum_layout(&self) -> Option<&'static str>

Gives the int type to use for the repr(C, int) enum layout