Module serde_derive_internals::attr

Structs

Attr

BoolAttr

Container
Represents container (e.g. struct) attribute information

Field
Represents field attribute information

Name

Variant
Represents variant attribute information

Enums

Default
Represents the default to use for a field when deserializing.

EnumTag
Styles of representing an enum.

Identifier
Whether this enum represents the fields of a struct or the variants of an enum.

RenameRule

Functions

borrowable_lifetimes

collect_lifetimes

decide_identifier

decide_tag

get_renames

get_ser_and_de

get_serde_meta_items

get_string_from_lit

get_where_predicates

is_cow

is_rptr

parse_lit_into_lifetimes

parse_lit_into_path

parse_lit_into_ty

parse_lit_into_where

Type Definitions

SerAndDe