Enum serde_derive_internals::attr::Default

pub enum Default {
    None,
    Default,
    Path(ExprPath),
}

Represents the default to use for a field when deserializing.

Variants
None

Field must always be specified because it does not have a default.

Default

The default is given by std::default::Default::default().

Path(ExprPath)

The default is given by this function.