Module serde_derive_internals::check

Functions

check
Cross-cutting checks that require looking at more than a single attrs object. Simpler checks should happen when parsing and building the attrs.

check_getter
Getters are only allowed inside structs (not enums) with the remote attribute.

check_identifier
The other attribute must be used at most once and it must be the last variant of an enum that has the field_identifier attribute.

check_variant_skip_attrs
Skip-(de)serializing attributes are not allowed on variants marked (de)serialize_with.