Module serde_derive_internals::attr

Structs

Container
Represents struct or enum attribute information.

Field
Represents field attribute information

Name

RenameAllRules

Variant
Represents variant attribute information

Enums

Default
Represents the default to use for a field when deserializing.

Identifier
Whether this enum represents the fields of a struct or the variants of an enum.

RenameRule
The different possible ways to change case of fields in a struct, or variants in an enum.

TagType
Styles of representing an enum.

Functions

get_serde_meta_items